Organization, naming improvements
